更改推荐视频以及其评论接口

+++ library_base/src/main/java/com/knight/kotlin/library_base/entity/EyeDailyListEntity.kt
@@ -88,7 +88,7 @@ data class EyeAuthorEntity(
     val expert: Boolean,
     val follow: EyeFollow,
     val icon: String,
-    val id: Int,
+    val id: Long,
     val ifPgc: Boolean,
     val latestReleaseTime: Long,
     val link: String,
@@ -125,10 +125,10 @@ data class EyeOwner(
 
 @Parcelize
 data class EyeConsumption(
-    val collectionCount: Int,
-    val realCollectionCount: Int,
-    val replyCount: Int,
-    val shareCount: Int
+    val collectionCount: Long,
+    val realCollectionCount: Long,
+    val replyCount: Long,
+    val shareCount: Long
 ): Parcelable
 @Parcelize
 data class EyeCover(
